requirements.txt
flwr==1.5.0
tensorflow-cpu==2.14.1
scikit-learn
pandas
matplotlib
jupyterlab
git+https://github.com/nepiskopos/py7zr.git
